Re-enabled the P/Invoke calls should be portable rule in FxCop and fixed the warnings that this produced.

+++ Controls/ExtListview.vb
@@ -41,7 +41,7 @@ Friend Class ExtListView : Inherits ListView
 
     ' API Declarations
     Private Declare Auto Function SendMessage Lib "user32" (ByVal hWnd As IntPtr, ByVal Msg As Integer, ByVal wParam As IntPtr, ByVal lParam As IntPtr) As IntPtr
-    Private Declare Unicode Function SetWindowTheme Lib "uxtheme" (ByVal hWnd As IntPtr, ByVal pszSubAppName As String, ByVal pszSubIdList As String) As IntPtr
+    Private Declare Unicode Function SetWindowTheme Lib "uxtheme" (ByVal hWnd As IntPtr, ByVal pszSubAppName As String, ByVal pszSubIdList As String) As Integer
 
     ' Data structure to store information about the controls
     Private Structure EmbeddedProgress
